Update lib/Core.php
Ok, the whole url was being escaped when it shouldn't have.

But we still needed it to escape whenever calling /v1/contacts/tags/<tag>.

So, the new code checks for this.  Couldn't think of any other service
that would need to be escaped before hand so just leaving it simple.